Add flow control for linked belts
Send message to stop flow if the output side buffer is filling up and
send another message to start flow again when the buffer is below a
pre-caculated limit. When an edge is deactivated all input flows are
stopped and when edge is activated all output flows are set to send
a start flow message when the content of the chest is empty.